add column header sort for tests table
