+++ src/pages/task/taskTabs/testsTable/TestsTableCore.tsx
@@ -1,14 +1,9 @@
 import { useEffect, useRef } from "react";
 import { useQuery } from "@apollo/client";
-import styled from "@emotion/styled";
-import Button from "@leafygreen-ui/button";
 import { Table } from "antd";
-import { ColumnProps } from "antd/es/table";
 import { SortOrder } from "antd/es/table/interface";
 import { useParams, useLocation } from "react-router-dom";
 import { useTaskAnalytics } from "analytics";
-import { Analytics } from "analytics/addPageAction";
-import Badge, { Variant } from "components/Badge";
 import { PageSizeSelector } from "components/PageSizeSelector";
 import { Pagination } from "components/Pagination";
 import { ResultCountLabel } from "components/ResultCountLabel";
@@ -17,13 +12,8 @@ import {
   TableControlOuterRow,
   TableControlInnerRow,
 } from "components/styles";
-import { WordBreak } from "components/Typography";
-import {
-  getLobsterTestLogUrl,
-  getUpdatedLobsterUrl,
-  isLogkeeperLink,
-} from "constants/externalResources";
 import { pollInterval } from "constants/index";
+import { testStatusesFilterTreeData } from "constants/test";
 import {
   TaskTestsQuery,
   TaskTestsQueryVariables,
@@ -33,11 +23,16 @@ import {
   TaskTestResult,
 } from "gql/generated/types";
 import { GET_TASK_TESTS } from "gql/queries";
-import { useUpdateURLQueryParams, useNetworkStatus } from "hooks";
-import { TestStatus, RequiredQueryParams, TableOnChange } from "types/task";
-import { queryString, url, string } from "utils";
+import {
+  useUpdateURLQueryParams,
+  useNetworkStatus,
+  useStatusesFilter,
+} from "hooks";
+import { RequiredQueryParams, TableOnChange } from "types/task";
+import { TestStatus } from "types/test";
+import { queryString, url } from "utils";
+import { getColumnsTemplate } from "./getColumnsTemplate";
 
-const { msToDuration } = string;
 const { getPageFromSearch, getLimitFromSearch } = url;
 const { parseQueryString, queryParamAsNumber } = queryString;
 
@@ -49,6 +44,8 @@ export const TestsTableCore: React.FC = () => {
   const { pathname, search } = useLocation();
   const updateQueryParams = useUpdateURLQueryParams();
   const taskAnalytics = useTaskAnalytics();
+  const sendFilterTestsEvent = (filterBy: string) =>
+    taskAnalytics.sendEvent({ name: "Filter Tests", filterBy });
 
   const queryVariables = getQueryVariables(search, resourceId);
   const { cat, dir, pageNum, limitNum } = queryVariables;
@@ -67,8 +64,25 @@ export const TestsTableCore: React.FC = () => {
       });
     }
   }, [pathname, updateQueryParams]); // eslint-disable-line react-hooks/exhaustive-deps
-  // Apply sorts to columns
-  const columns = getColumnsTemplate(taskAnalytics).map((column) => ({
+
+  const statusesFilter = useStatusesFilter(
+    RequiredQueryParams.Statuses,
+    false,
+    sendFilterTestsEvent
+  );
+
+  const statusSelectorProps = {
+    state: statusesFilter.inputValue,
+    tData: testStatusesFilterTreeData,
+    onChange: statusesFilter.setInputValue,
+    onReset: statusesFilter.reset,
+    onFilter: statusesFilter.submitInputValue,
+  };
+
+  const columns = getColumnsTemplate({
+    taskAnalytics,
+    statusSelectorProps,
+  }).map((column) => ({
     ...column,
     ...(column.key === cat && {
       sortOrder: (dir === SortDirection.Asc
@@ -148,24 +162,7 @@ export const TestsTableCore: React.FC = () => {
   );
 };
 
-const statusToBadgeColor = {
-  [TestStatus.Pass]: Variant.Green,
-  [TestStatus.Fail]: Variant.Red,
-  [TestStatus.SilentFail]: Variant.Blue,
-  [TestStatus.Skip]: Variant.Yellow,
-};
-const statusCopy = {
-  [TestStatus.Pass]: "Pass",
-  [TestStatus.Fail]: "Fail",
-  [TestStatus.Skip]: "Skip",
-  [TestStatus.SilentFail]: "Silent Fail",
-};
-
 export const rowKey = ({ id }: { id: string }): string => id;
-
-const ButtonWrapper = styled("span")`
-  margin-right: 8px;
-`;
 
 const getQueryVariables = (
   search: string,
@@ -204,134 +201,3 @@ const getQueryVariables = (
     execution: queryParamAsNumber(execution),
   };
 };
-
-const getColumnsTemplate = (
-  taskAnalytics: Analytics<
-    | { name: "Click Logs Lobster Button" }
-    | { name: "Click Logs HTML Button" }
-    | { name: "Click Logs Raw Button" }
-  >
-): ColumnProps<TestResult>[] => [
-  {
-    title: <span data-cy="name-column">Name</span>,
-    dataIndex: "testFile",
-    key: TestSortCategory.TestName,
-    width: "40%",
-    render: (name, { displayTestName }) => (
-      <WordBreak>{displayTestName || name}</WordBreak>
-    ),
-    sorter: true,
-  },
-  {
-    title: <span data-cy="status-column">Status</span>,
-    dataIndex: "status",
-    key: TestSortCategory.Status,
-    sorter: true,
-    render: (status: string): JSX.Element => (
-      <span>
-        <Badge
-          variant={statusToBadgeColor[status] || Variant.LightGray}
-          key={status}
-        >
-          {statusCopy[status] || ""}
-        </Badge>
-      </span>
-    ),
-  },
-  {
-    title: <span data-cy="base-status-column">Base Status</span>,
-    dataIndex: "baseStatus",
-    key: TestSortCategory.BaseStatus,
-    sorter: true,
-    render: (status: string): JSX.Element => (
-      <span>
-        <Badge
-          variant={statusToBadgeColor[status] || Variant.LightGray}
-          key={status}
-        >
-          {statusCopy[status] || ""}
-        </Badge>
-      </span>
-    ),
-  },
-  {
-    title: <span data-cy="time-column">Time</span>,
-    dataIndex: "duration",
-    key: TestSortCategory.Duration,
-    sorter: true,
-    render: (text: number): string => {
-      const ms = text * 1000;
-      return msToDuration(Math.trunc(ms));
-    },
-  },
-  {
-    title: <span data-cy="logs-column">Logs</span>,
-    width: 230,
-    dataIndex: "logs",
-    key: "logs",
-    sorter: false,
-    render: (a, b): JSX.Element => {
-      const { execution, lineNum, taskId, id } = b || {};
-      const { htmlDisplayURL, rawDisplayURL } = b?.logs ?? {};
-      const hasLobsterLink = isLogkeeperLink(htmlDisplayURL);
-      const lobsterLink = hasLobsterLink
-        ? getUpdatedLobsterUrl(htmlDisplayURL)
-        : getLobsterTestLogUrl({ taskId, execution, testId: id, lineNum });
-
-      return (
-        <>
-          {lobsterLink && (
-            <ButtonWrapper>
-              <Button
-                data-cy="test-table-lobster-btn"
-                size="small"
-                target="_blank"
-                variant="default"
-                href={lobsterLink}
-                onClick={() =>
-                  taskAnalytics.sendEvent({
-                    name: "Click Logs Lobster Button",
-                  })
-                }
-              >
-                Lobster
-              </Button>
-            </ButtonWrapper>
-          )}
-          {!hasLobsterLink && (
-            <ButtonWrapper>
-              <Button
-                data-cy="test-table-html-btn"
-                size="small"
-                target="_blank"
-                variant="default"
-                href={htmlDisplayURL}
-                onClick={() =>
-                  taskAnalytics.sendEvent({
-                    name: "Click Logs HTML Button",
-                  })
-                }
-              >
-                HTML
-              </Button>
-            </ButtonWrapper>
-          )}
-          {rawDisplayURL && (
-            <Button
-              data-cy="test-table-raw-btn"
-              size="small"
-              target="_blank"
-              variant="default"
-              href={rawDisplayURL}
-              onClick={() =>
-                taskAnalytics.sendEvent({ name: "Click Logs Raw Button" })
-              }
-            >
-              Raw
-            </Button>
-          )}
-        </>
-      );
-    },
-  },
-];
