+++ src/types/test.ts
@@ -1,0 +1,7 @@
+export enum TestStatus {
+  Fail = "fail",
+  Skip = "skip",
+  SilentFail = "silentfail",
+  Pass = "pass",
+  All = "all",
+}
